backend/requirements.txt
Django>=5.1
djangorestframework>=3.14.0
django-cors-headers>=3.7.0
python-dotenv>=1.0.0
bcrypt>=3.2.0
black>=21.12b0